03-microservices-scaling/microservices/auth-service/requirements.txt
flask
flask-sqlalchemy
flask-migrate
pymysql
pyjwt
python-dotenv
gunicorn